refactor(how-why): factor out deepDependencies, zenhubTools, ...
... and components
